+++ crates/ferrmion-core/src/optimise/encoding/mod.rs
@@ -1,0 +1,2 @@
+mod anneal;
+pub use anneal::*;