+++ crates/ferrmion-core/src/optimise/hamiltonian/mod.rs
@@ -1,0 +1,2 @@
+mod cliffordheuristic;
+pub use cliffordheuristic::*;