+++ crates/ferrmion-core/src/optimise/ternarytree/mod.rs
@@ -1,0 +1,4 @@
+mod hatt;
+pub use hatt::*;
+mod topphatt;
+pub use topphatt::*;